moved customization variables

+++ Server Only/R2Northstar/mods/Northstar.CustomServers/mod/scripts/vscripts/gamemodes/_gamemode_ffa.nut
@@ -1,5 +1,19 @@
 global function FFA_Init
 
+struct {
+	//// CHANGE ME
+	string score_leader_highlight = "enemy_boss_bounty" // highlight effect applied to person in 1st place
+	string offhand_weapon = "mp_weapon_thermite_grenade" // offhand weapon
+
+	bool reset_pulse_blade_cooldown_on_pulse_blade_kill = true
+
+	int wme_kill_value = 10
+	int offhand_kill_value = 10
+	int reset_kill_value = 5
+	int melee_kill_value = 5
+	//// STOP CHANGING ME
+} file
+
 void function FFA_Init()
 {
 	SetShouldUseRoundWinningKillReplay( true )
@@ -21,22 +35,33 @@ void function OnPlayerKilled( entity victim, entity attacker, var damageInfo )
 		SetRoundWinningKillReplayAttacker(attacker)
 		if ( DamageInfo_GetDamageSourceIdentifier( damageInfo ) == eDamageSourceId.mp_weapon_grenade_sonar || DamageInfo_GetDamageSourceIdentifier( damageInfo ) == eDamageSourceId.human_execution)
 		{
-			attacker.TakeWeaponNow( "mp_weapon_grenade_sonar" ) // resets cooldown if you kill with it (i think)
-			attacker.GiveOffhandWeapon( "mp_weapon_grenade_sonar", OFFHAND_LEFT )
-
+			if (file.reset_pulse_blade_cooldown_on_pulse_blade_kill) 
+			{
+				attacker.TakeWeaponNow( "mp_weapon_grenade_sonar" ) // resets cooldown if you kill with it
+				attacker.GiveOffhandWeapon( "mp_weapon_grenade_sonar", OFFHAND_LEFT )
+			}
+			
 			EmitSoundOnEntityOnlyToPlayer( attacker, attacker, "UI_CTF_3P_TeamGrabFlag" )
 			bankrupt(victim)
 
-			AddTeamScore( attacker.GetTeam(), 5 )
-			attacker.AddToPlayerGameStat( PGS_ASSAULT_SCORE, 5 )
+			AddTeamScore( attacker.GetTeam(), file.reset_kill_value )
+			attacker.AddToPlayerGameStat( PGS_ASSAULT_SCORE, file.reset_kill_value )
 			attacker.AddToPlayerGameStat( PGS_TITAN_KILLS, 1 )
-		} else if ( DamageInfo_GetDamageSourceIdentifier( damageInfo ) == eDamageSourceId.melee_pilot_emptyhanded ) 
+		} 
+		else if ( DamageInfo_GetDamageSourceIdentifier( damageInfo ) == eDamageSourceId.melee_pilot_emptyhanded ) 
 		{
-			AddTeamScore( attacker.GetTeam(), 5 )
-			attacker.AddToPlayerGameStat( PGS_ASSAULT_SCORE, 5 )
-		} else {
-			AddTeamScore( attacker.GetTeam(), 10 )
-			attacker.AddToPlayerGameStat( PGS_ASSAULT_SCORE, 10 )
+			AddTeamScore( attacker.GetTeam(), file.melee_kill_value )
+			attacker.AddToPlayerGameStat( PGS_ASSAULT_SCORE, file.melee_kill_value )
+		} 
+		else if ( DamageInfo_GetDamageSourceIdentifier( damageInfo ) == eDamageSourceId.mp_weapon_wingman_n ) 
+		{
+			AddTeamScore( attacker.GetTeam(), file.wme_kill_value )
+			attacker.AddToPlayerGameStat( PGS_ASSAULT_SCORE, file.wme_kill_value )
+		}
+		else 
+		{
+			AddTeamScore( attacker.GetTeam(), file.offhand_kill_value )
+			attacker.AddToPlayerGameStat( PGS_ASSAULT_SCORE, file.offhand_kill_value )
 		}
 
         if (attacker == GetWinningPlayer())
@@ -70,7 +95,7 @@ void function OnPlayerRespawned( entity player )
 	array<string> mods = ["sns", "pas_fast_ads", "tactical_cdr_on_kill", "pas_run_and_gun", "pas_fast_swap"]
 	player.GiveWeapon( "mp_weapon_wingman_n", mods)
 	player.GiveOffhandWeapon( "melee_pilot_emptyhanded", OFFHAND_MELEE )
-	player.GiveOffhandWeapon( "mp_weapon_thermite_grenade", OFFHAND_RIGHT )
+	player.GiveOffhandWeapon( file.offhand_weapon, OFFHAND_RIGHT )
 	player.GiveOffhandWeapon( "mp_weapon_grenade_sonar", OFFHAND_LEFT )
 
     if (player == GetWinningPlayer())
@@ -106,5 +131,5 @@ entity function GetWinningPlayer()
 void function SetHighlight(entity player) {
     foreach ( entity player in GetPlayerArray() )
         Highlight_ClearEnemyHighlight(player)
-    Highlight_SetEnemyHighlight( player, "enemy_boss_bounty" )
+    Highlight_SetEnemyHighlight( player, file.score_leader_highlight )
 }
